Extract first-pass ingestion into cgt_calc/ingestion.py
Fourth step of splitting main.py into phase modules. Moves the
first pass - the add_* handlers, gift resolution, stock split
planning, spin-off handling, the transaction dispatcher with its
balance check, and the first-pass report - into a
TransactionIngester that takes the shared CalculatorState plus the
converters and the income processor it reads. The module-level
or-fail, price-rounding, gift-matching and split helpers move with
their only callers. The calculator keeps
convert_to_hmrc_transactions and first_pass_report delegates for
external callers, and one caplog filter follows the spouse-transfer
warning to the cgt_calc.ingestion logger. The split replay methods
are second-pass matching and stay put. Pure move, no behavior
change.
